ver. 8.2

+++ app/Http/Controllers/HomeController.php
@@ -5,13 +5,12 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Advt;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\DB;
 
 class HomeController extends Controller
 {
     public function index()
     {
-        $advts = Advt::orderBy('id','desc')->paginate(5);
+        $advts = Advt::orderBy('id', 'desc')->paginate(5);
         return view('index', ['advts' => $advts]);
     }
 
@@ -19,14 +18,16 @@ class HomeController extends Controller
     {
         $request = request();
 
-        $this->validate($request,[
+        $this->validate($request, [
             'title' => ['required'],
-            'description'=> ['required']
+            'description' => ['required']
         ]);
 
         $id = $request->route()->parameter('id');
         $advt['advt'] = Advt::find($id);
-
+        if (isset($advt['advt']) && $advt['advt']->user_id !== Auth::id()) {
+            return redirect()->route('home');
+        }
         $advt['advt']->update([
             'title' => $request->get('title'),
             'description' => $request->get('description')
@@ -50,9 +51,11 @@ class HomeController extends Controller
             $advt['advt'] = Advt::find($id);
             $advt['action'] = $id;
             $advt['buttonName'] = 'Save';
+            //return redirect()->route('home');
         } else {
-            $advt['action'] = 'create';
+            $advt['action'] = '';
             $advt['buttonName'] = 'Create';
+            //return redirect()->route('home');
         }
         return view('edit.form', ['advt' => $advt]);
 
@@ -60,9 +63,9 @@ class HomeController extends Controller
 
     public function create(Request $request)
     {
-        $this->validate($request,[
+        $this->validate($request, [
             'title' => ['required'],
-            'description'=> ['required']
+            'description' => ['required']
         ]);
 
         $advt = Advt::create([
@@ -82,6 +85,9 @@ class HomeController extends Controller
     public function delete($id)
     {
         $advt = Advt::find($id);
+        if (isset($advt) && $advt->user_id !== Auth::id()) {
+            return redirect()->route('home');
+        }
         $advt->delete();
         return redirect('/');
     }
